Struct tmc_langs_core::TmcCore

pub struct TmcCore { /* fields omitted */ }

A struct for interacting with the TestMyCode service, including authentication

Methods

impl TmcCore

pub fn user(&self, user_id: usize) -> Result<User, CoreError>

pub fn user_current(&self) -> Result<User, CoreError>

pub fn basic_info_by_usernames(&self) -> Result<Vec<User>, CoreError>

pub fn basic_info_by_emails(&self) -> Result<Vec<User>, CoreError>

pub fn course(&self, course_id: usize) -> Result<NuCourse, CoreError>

pub fn course_by_name(
&self,
organization_slug: &str,
course_name: &str
) -> Result<NuCourse, CoreError>

pub fn course_points(&self, course_id: usize) -> Result<(), CoreError>

pub fn exercise_points(
&self,
course_id: usize,
exercise_name: &str
) -> Result<Vec<NuExercisePoint>, CoreError>

pub fn exercise_points_for_user(
&self,
course_id: usize,
exercise_name: &str,
user_id: usize
) -> Result<Vec<NuExercisePoint>, CoreError>

pub fn exercise_points_for_current_user(
&self,
course_id: usize,
exercise_name: &str
) -> Result<Vec<NuExercisePoint>, CoreError>

pub fn course_points_for_user(
&self,
course_id: usize,
user_id: usize
) -> Result<Vec<NuExercisePoint>, CoreError>

pub fn course_points_for_current_user(
&self,
course_id: usize
) -> Result<Vec<NuExercisePoint>, CoreError>

pub fn course_points_by_name(
&self,
organization_slug: &str,
course_name: &str
) -> Result<Vec<NuExercisePoint>, CoreError>

pub fn eligible_students(
&self,
organization_slug: &str,
course_name: &str
) -> Result<(), CoreError>

pub fn exercise_points_by_name(
&self,
organization_slug: &str,
course_name: &str,
exercise_name: &str
) -> Result<(), CoreError>

pub fn exercise_points_by_name_for_current_user(
&self,
organization_slug: &str,
course_name: &str,
exercise_name: &str
) -> Result<Vec<NuExercisePoint>, CoreError>

pub fn exercise_points_by_name_for_user(
&self,
organization_slug: &str,
course_name: &str,
exercise_name: &str,
user_id: usize
) -> Result<Vec<NuExercisePoint>, CoreError>

pub fn course_points_by_name_for_user(
&self,
organization_slug: &str,
course_name: &str,
user_id: usize
) -> Result<Vec<NuExercisePoint>, CoreError>

pub fn course_points_by_name_for_current_user(
&self,
organization_slug: &str,
course_name: &str
) -> Result<Vec<NuExercisePoint>, CoreError>

pub fn course_submissions(
&self,
course_id: usize
) -> Result<Vec<Submission>, CoreError>

pub fn course_submissions_in_last_hour(
&self,
course_id: usize
) -> Result<Vec<Submission>, CoreError>

pub fn course_submissions_for_user(
&self,
course_id: usize,
user_id: usize
) -> Result<Vec<Submission>, CoreError>

pub fn course_submissions_for_current_user(
&self,
course_id: usize
) -> Result<Vec<Submission>, CoreError>

pub fn exercise_submissions_for_user(
&self,
exercise_id: usize,
user_id: usize
) -> Result<Vec<Submission>, CoreError>

pub fn exercise_submissions_for_current_user(
&self,
exercise_id: usize
) -> Result<Vec<Submission>, CoreError>

pub fn exercise_submissions_by_name(
&self,
organization_slug: &str,
course_name: &str
) -> Result<Vec<Submission>, CoreError>

pub fn exercise_submissions_by_name_for_user(
&self,
organization_slug: &str,
course_name: &str,
user_id: usize
) -> Result<Vec<Submission>, CoreError>

pub fn exercise_submissions_by_name_for_currrent_user(
&self,
organization_slug: &str,
course_name: &str
) -> Result<Vec<Submission>, CoreError>

pub fn exercises(
&self,
course_id: usize
) -> Result<Vec<CourseExercise>, CoreError>

pub fn exercises_by_name(
&self,
organization_slug: &str,
course_name: &str
) -> Result<Vec<NuCourseExercise>, CoreError>

pub fn download_exercise_by_name(
&self,
organization_slug: &str,
course_name: &str,
exercise_name: &str,
target: &Path
) -> Result<(), CoreError>

pub fn organizations(&self) -> Result<Vec<Organization>, CoreError>

pub fn organization(
&self,
organization_slug: &str
) -> Result<Organization, CoreError>

pub fn core_course(&self, course_id: usize) -> Result<CourseDetails, CoreError>

pub fn reviews(&self, course_id: usize) -> Result<Vec<Review>, CoreError>

pub fn review(
&self,
course_id: usize,
review_id: usize
) -> Result<Vec<Review>, CoreError>

pub fn unlock(&self, course_id: usize) -> Result<(), CoreError>

pub fn download_exercise(
&self,
exercise_id: usize,
target: &Path
) -> Result<(), CoreError>

pub fn core_exercise(
&self,
exercise_id: usize
) -> Result<ExerciseDetails, CoreError>

pub fn download_solution(
&self,
exercise_id: usize,
target: &Path
) -> Result<(), CoreError>

pub fn post_submission(
&self,
exercise_id: usize,
submission: &Path
) -> Result<NewSubmission, CoreError>

pub fn organization_courses(
&self,
organization_slug: &str
) -> Result<Vec<Course>, CoreError>

pub fn download_submission(
&self,
submission_id: usize,
target: &Path
) -> Result<(), CoreError>

pub fn post_feedback(
&self,
submission_id: usize,
feedback: Vec<FeedbackAnswer>
) -> Result<SubmissionFeedbackResponse, CoreError>

pub fn post_review(
&self,
submission_id: usize,
review_body: &str,
review_points: &str
) -> Result<(), CoreError>

impl TmcCore

pub fn new(
config_dir: PathBuf,
root_url: &'static str
) -> Result<Self, CoreError>

pub fn new_in_config(root_url: &'static str) -> Result<Self, CoreError>

pub fn authenticate(
&mut self,
client_name: &str,
username: String,
password: String
) -> Result<(), CoreError>

Attempts to log in with the given credentials, returns an error if an authentication token is already present. Username can be the user's username or email.

pub fn unauthenticate(&mut self)

pub fn is_authenticated(&self) -> bool

pub fn get_user_info(&self) -> Result<User, CoreError>

pub fn get_organizations(&self) -> Result<Vec<Organization>, CoreError>

Returns a list of organizations.

pub fn get_organization(
&self,
organization_slug: &str
) -> Result<Organization, CoreError>

Returns an organization.

pub fn get_courses(
&self,
organization_slug: &str
) -> Result<Vec<Course>, CoreError>

pub fn get_course_details(
&self,
course_id: usize
) -> Result<CourseDetails, CoreError>

pub fn get_course_exercises(
&self,
course_id: usize
) -> Result<Vec<CourseExercise>, CoreError>

pub fn get_exercise_details(
&self,
exercise_id: usize
) -> Result<ExerciseDetails, CoreError>

pub fn download_exercise_old(
&self,
exercise_id: u32,
target: &Path
) -> Result<(), CoreError>

pub fn get_submissions(
&self,
course_id: usize
) -> Result<Vec<Submission>, CoreError>

pub fn download_submission_old(
&self,
submission_id: usize,
target: &Path
) -> Result<(), CoreError>

pub fn submit_exercise(
&self,
exercise_id: usize,
exercise_path: &Path
) -> Result<NewSubmission, CoreError>

pub fn check_submission(
&self,
submission_url: &str
) -> Result<SubmissionProcessingStatus, CoreError>

pub fn submit_feedback(
&self,
feedback_url: &str,
feedback_answers: Vec<FeedbackAnswer>
) -> Result<SubmissionFeedbackResponse, CoreError>

pub fn download_model_solution(
&self,
exercise_id: usize,
target: &Path
) -> Result<(), CoreError>

pub fn get_unread_reviews(
&self,
course_id: usize
) -> Result<Vec<Review>, CoreError>

pub fn get_updateable_exercises(
&self,
course_id: usize
) -> Result<(), CoreError>