Struct tmc_langs_core::TmcCore

pub struct TmcCore { /* fields omitted */ }

A struct for interacting with the TestMyCode service, including authentication

Methods

impl TmcCore

pub fn download_from(&self, url: Url, target: &Path) -> Result<(), CoreError>

impl TmcCore

pub fn new(config_dir: PathBuf, root_url: String) -> Result<Self, CoreError>

pub fn new_in_config(root_url: String) -> Result<Self, CoreError>

pub fn authenticate(
&mut self,
client_name: &str,
email: String,
password: String
) -> Result<(), CoreError>

Attempts to log in with the given credentials, returns an error if an authentication token is already present. Username can be the user's username or email.

pub fn get_organizations(&self) -> Result<Vec<Organization>, CoreError>

pub fn send_diagnostics(&self)

pub fn download_or_update_exercises(
&self,
exercises: Vec<(usize, &Path)>
) -> Result<(), CoreError>

pub fn get_course_details(
&self,
course_id: usize
) -> Result<CourseDetails, CoreError>

pub fn list_courses(
&self,
organization_slug: &str
) -> Result<Vec<Course>, CoreError>

pub fn paste_with_comment(
&self,
submission_url: Url,
submission_path: &Path,
paste_message: String
) -> Result<NewSubmission, CoreError>

pub fn run_checkstyle(
&self,
path: &Path,
locale: Language
) -> Result<Option<ValidationResult>, CoreError>

pub fn run_tests(&self, path: &Path) -> Result<RunResult, CoreError>

pub fn send_feedback(
&self,
feedback_url: Url,
feedback: Vec<FeedbackAnswer>
) -> Result<SubmissionFeedbackResponse, CoreError>

pub fn send_snapshot_events(&self)

pub fn submit(
&self,
submission_url: Url,
submission_path: &Path
) -> Result<NewSubmission, CoreError>

pub fn get_exercise_updates(
&self,
course_id: usize,
checksums: HashMap<usize, String>
) -> Result<UpdateResult, CoreError>

Fetches the course's exercises from the server, and finds new or updated exercises. If an exercise's id is not found in the checksum map, it is considered new. If an id is found, it is compared to the current one. If they are different, it is considered updated.

pub fn mark_review_as_read(
&self,
review_update_url: String
) -> Result<(), CoreError>

pub fn get_unread_reviews(
&self,
course_id: usize
) -> Result<Vec<Review>, CoreError>

pub fn request_code_review(
&self,
submission_url: Url,
submission_path: &Path,
message_for_reviewer: String
) -> Result<NewSubmission, CoreError>

pub fn download_model_solution(
&self,
solution_download_url: Url,
target: &Path
) -> Result<(), CoreError>

pub fn check_submission(
&self,
submission_url: &str
) -> Result<SubmissionProcessingStatus, CoreError>