Struct tmc_langs_core::TmcCore

pub struct TmcCore(_);

A struct for interacting with the TestMyCode service, including authentication.

Implementations

impl TmcCore

pub fn get_json_from_url<T: DeserializeOwned>(
&self,
url: Url,
params: &[(String, String)]
) -> Result<T, CoreError>

impl TmcCore

pub fn new(
config_dir: PathBuf,
root_url: String,
client_name: String,
client_version: String
) -> Result<Self, CoreError>

Creates a new TmcCore with the given config directory and root URL.

Errors

This function will return an error if parsing the root_url fails.

Examples

use tmc_langs_core::TmcCore;
use std::path::PathBuf;

let core = TmcCore::new(PathBuf::from("./config"), "https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();

pub fn new_in_config(
root_url: String,
client_name: String,
client_version: String
) -> Result<Self, CoreError>

Creates a new TmcCore with the given root URL. The config directory is set according to dirs::cache_dir.

Errors

This function will return an error if parsing the root_url fails, or if fetching the cache directory fails (see dirs::cache_dir()).

Examples

use tmc_langs_core::TmcCore;

let core = TmcCore::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_ver".to_string()).unwrap();

pub fn set_token(&mut self, token: Token) -> Result<(), CoreError>

pub fn set_progress_reporter(
&mut self,
progress_reporter: ProgressReporter<'static, CoreUpdateData>
) -> Result<(), CoreError>

pub fn increment_progress_steps(&self)

pub fn authenticate(
&mut self,
client_name: &str,
email: String,
password: String
) -> Result<Token, CoreError>

Attempts to log in with the given credentials, returns an error if an authentication token is already present. Username can be the user's username or email.

Errors

This function will return an error if the core has already been authenticated, if the client_name is malformed and leads to a malformed URL, or if there is some error during the token exchange (see oauth2::Client::excange_password).

Examples

use tmc_langs_core::TmcCore;

let mut core = TmcCore::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
core.authenticate("client", "user".to_string(), "pass".to_string()).unwrap();

pub fn get_organizations(&self) -> Result<Vec<Organization>, CoreError>

Fetches all organizations.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error.

pub fn get_organization(
&self,
organization_slug: &str
) -> Result<Organization, CoreError>

Fetches an organization.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error.

pub fn download_or_update_exercises(
&self,
exercises: Vec<(usize, PathBuf)>
) -> Result<(), CoreError>

Downloads the given exercises. Overwrites existing exercises if they exist.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error. The method extracts zip archives, which may fail.

Examples

use tmc_langs_core::TmcCore;
use std::path::PathBuf;

let core = TmcCore::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
// authenticate
core.download_or_update_exercises(vec![
    (1234, PathBuf::from("./exercises/1234")),
    (2345, PathBuf::from("./exercises/2345")),
]);

pub fn get_course_details(
&self,
course_id: usize
) -> Result<CourseDetails, CoreError>

Fetches the course's information.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error.

Examples

use tmc_langs_core::TmcCore;

let core = TmcCore::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
// authenticate
let course_details = core.get_course_details(600).unwrap();

pub fn get_exercise_details(
&self,
exercise_id: usize
) -> Result<ExerciseDetails, CoreError>

pub fn get_exercises_details(
&self,
exercise_ids: Vec<usize>
) -> Result<Vec<ExercisesDetails>, CoreError>

pub fn get_course_submissions(
&self,
course_id: usize
) -> Result<Vec<Submission>, CoreError>

pub fn list_courses(
&self,
organization_slug: &str
) -> Result<Vec<Course>, CoreError>

Fetches all courses under the given organization.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error.

Examples

use tmc_langs_core::TmcCore;

let core = TmcCore::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
// authenticate
let courses = core.list_courses("hy").unwrap();

pub fn get_course(&self, course_id: usize) -> Result<CourseData, CoreError>

pub fn get_course_exercises(
&self,
course_id: usize
) -> Result<Vec<CourseExercise>, CoreError>

pub fn paste(
&self,
submission_url: Url,
submission_path: &Path,
paste_message: Option<String>,
locale: Option<Language>
) -> Result<NewSubmission, CoreError>

Sends the given submission as a paste.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error.

Examples

use tmc_langs_core::{TmcCore, Language};
use url::Url;
use std::path::Path;

let core = TmcCore::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
// authenticate
let course_details = core.get_course_details(600).unwrap();
let submission_url = &course_details.exercises[0].return_url;
let submission_url = Url::parse(&submission_url).unwrap();
let new_submission = core.paste(
    submission_url,
    Path::new("./exercises/python/123"),
    Some("my python solution".to_string()),
    Some(Language::Eng)).unwrap();

pub fn run_checkstyle(
&self,
path: &Path,
locale: Language
) -> Result<Option<ValidationResult>, CoreError>

Checks the coding style for the project.

Errors

Returns an error if no matching language plugin for the project is found, or if the plugin returns an error while trying to run the style check.

Examples

use tmc_langs_core::{TmcCore, Language};
use std::path::Path;

let core = TmcCore::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
// authenticate
let validation_result = core.run_checkstyle(Path::new("./exercises/python/123"), Language::Eng).unwrap();
match validation_result {
    Some(validation_result) => if let Some(validation_errors) = validation_result.validation_errors {
        println!("found validation errors: {:?}", validation_errors);
    } else {
        println!("no errors");
    }
    None => println!("no style checks"),
}

pub fn run_tests(
&self,
path: &Path,
warnings: &mut Vec<Error>
) -> Result<RunResult, CoreError>

Runs tests for the project.

Errors

Returns an error if no matching language plugin for the project is found, or if the plugin returns an error while trying to run the tests.

pub fn send_feedback(
&self,
feedback_url: Url,
feedback: Vec<FeedbackAnswer>
) -> Result<SubmissionFeedbackResponse, CoreError>

Sends feedback.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error.

pub fn submit(
&self,
submission_url: Url,
submission_path: &Path,
locale: Option<Language>
) -> Result<NewSubmission, CoreError>

Sends the submission to the server.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error. The method compresses the submission and writes it into a temporary archive, which may fail.

pub fn reset(
&self,
exercise_id: usize,
exercise_path: PathBuf
) -> Result<(), CoreError>

pub fn download_old_submission(
&self,
submission_id: usize,
target: &Path
) -> Result<(), CoreError>

pub fn get_exercise_submissions_for_current_user(
&self,
exercise_id: usize
) -> Result<Vec<Submission>, CoreError>

pub fn wait_for_submission(
&self,
submission_url: &str
) -> Result<SubmissionFinished, CoreError>

pub fn get_exercise_updates(
&self,
course_id: usize,
checksums: HashMap<usize, String>
) -> Result<UpdateResult, CoreError>

Fetches the course's exercises from the server, and finds new or updated exercises. If an exercise's id is not found in the checksum map, it is considered new. If an id is found, it is compared to the current one. If they are different, it is considered updated.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error.

Examples

use tmc_langs_core::TmcCore;

let core = TmcCore::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
// authenticate
let mut checksums = std::collections::HashMap::new();
checksums.insert(1234, "exercisechecksum".to_string());
let update_result = core.get_exercise_updates(600, checksums).unwrap();

pub fn mark_review_as_read(
&self,
review_update_url: String
) -> Result<(), CoreError>

Mark the review as read on the server.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error.

pub fn get_unread_reviews(
&self,
reviews_url: Url
) -> Result<Vec<Review>, CoreError>

Fetches all reviews.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error.

pub fn request_code_review(
&self,
submission_url: Url,
submission_path: &Path,
message_for_reviewer: String,
locale: Option<Language>
) -> Result<NewSubmission, CoreError>

Request code review.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error. The method compresses the project and writes a temporary archive, which may fail.

pub fn download_model_solution(
&self,
solution_download_url: Url,
target: &Path
) -> Result<(), CoreError>

Downloads the model solution from the given url.

Errors

Returns an error if there's some problem reaching the API, or if the API returns an error. The method extracts the downloaded model solution archive, which may fail.

pub fn check_submission(
&self,
submission_url: &str
) -> Result<SubmissionProcessingStatus, CoreError>

Checks the status of a submission on the server.

Errors

Returns an error if the core has not been authenticated, or if there's some problem reaching the API, or if the API returns an error.